# pipeline/api-specs/openapi.yaml
components:
  schemas:
...
    CreateWikiSchemaDraftRequest:
      properties:
        name:
          default: default
          minLength: 1
          title: Name
          type: string
        raw_markdown:
          minLength: 1
          title: Raw Markdown
          type: string
        user_id:
          minLength: 1
          title: User Id
          type: string
        workspace_id:
          minLength: 1
          title: Workspace Id
          type: string
      required:
      - raw_markdown
      - workspace_id
      - user_id
      title: CreateWikiSchemaDraftRequest
      type: object
    CreateWikiSchemaDraftResponse:
      properties:
        wiki_schema:
          $ref: '#/components/schemas/WikiSchemaResponse'
      required:
      - wiki_schema
      title: CreateWikiSchemaDraftResponse
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    SchemaFragmentsResponse:
      properties:
        concept_markdown:
          title: Concept Markdown
          type: string
        edit_markdown:
          title: Edit Markdown
          type: string
        global_markdown:
          title: Global Markdown
          type: string
        ingest_markdown:
          title: Ingest Markdown
          type: string
        query_markdown:
          title: Query Markdown
          type: string
        template_markdown:
          title: Template Markdown
          type: string
      required:
      - global_markdown
      - query_markdown
      - ingest_markdown
      - edit_markdown
      - concept_markdown
      - template_markdown
      title: SchemaFragmentsResponse
      type: object
    SchemaIssueResponse:
      properties:
        category:
          title: Category
          type: string
        reason:
          title: Reason
          type: string
        section:
          anyOf:
          - type: string
          - type: 'null'
          title: Section
        severity:
          enum:
          - blocked
          - unclear
          title: Severity
          type: string
        text:
          title: Text
          type: string
      required:
      - severity
      - category
      - text
      - reason
      title: SchemaIssueResponse
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
    WikiSchemaPreviewRequest:
      properties:
        raw_markdown:
          minLength: 1
          title: Raw Markdown
          type: string
      required:
      - raw_markdown
      title: WikiSchemaPreviewRequest
      type: object
    WikiSchemaPreviewResponse:
      properties:
        fragments:
          $ref: '#/components/schemas/SchemaFragmentsResponse'
        has_blocked_issues:
          title: Has Blocked Issues
          type: boolean
        issues:
          items:
            $ref: '#/components/schemas/SchemaIssueResponse'
          title: Issues
          type: array
        preview_markdown:
          title: Preview Markdown
          type: string
      required:
      - fragments
      - issues
      - preview_markdown
      - has_blocked_issues
      title: WikiSchemaPreviewResponse
      type: object
    WikiSchemaResponse:
      properties:
        activated_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Activated At
        created_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Created At
        fragments:
          $ref: '#/components/schemas/SchemaFragmentsResponse'
        has_blocked_issues:
          title: Has Blocked Issues
          type: boolean
        id:
          title: Id
          type: string
        issues:
          items:
            $ref: '#/components/schemas/SchemaIssueResponse'
          title: Issues
          type: array
        name:
          title: Name
          type: string
        preview_markdown:
          title: Preview Markdown
          type: string
        raw_markdown:
          title: Raw Markdown
          type: string
        schema_version:
          title: Schema Version
          type: string
        status:
          title: Status
          type: string
        updated_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Updated At
        user_id:
          title: User Id
          type: string
        workspace_id:
          title: Workspace Id
          type: string
      required:
      - id
      - workspace_id
      - user_id
      - name
      - raw_markdown
      - fragments
      - issues
      - preview_markdown
      - has_blocked_issues
      - status
      - schema_version
      title: WikiSchemaResponse
      type: object
info:
  title: Fruition Pipeline Lab API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /wiki-schema/active:
    get:
      operationId: get_active_wiki_schema_wiki_schema_active_get
      parameters:
      - in: query
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      - in: query
        name: user_id
        required: true
        schema:
          title: User Id
          type: string
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/WikiSchemaResponse'
                - type: 'null'
                title: Response Get Active Wiki Schema Wiki Schema Active Get
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Active Wiki Schema
      tags:
      - wiki-schema
  /wiki-schema/drafts:
    post:
      operationId: create_wiki_schema_draft_wiki_schema_drafts_post
      parameters:
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWikiSchemaDraftRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateWikiSchemaDraftResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Wiki Schema Draft
      tags:
      - wiki-schema
  /wiki-schema/preview:
    post:
      operationId: preview_wiki_schema_wiki_schema_preview_post
      parameters:
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiSchemaPreviewRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WikiSchemaPreviewResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Preview Wiki Schema
      tags:
      - wiki-schema
  /wiki-schema/{schema_id}/activate:
    post:
      operationId: activate_wiki_schema_wiki_schema__schema_id__activate_post
      parameters:
      - in: path
        name: schema_id
        required: true
        schema:
          title: Schema Id
          type: string
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WikiSchemaResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Activate Wiki Schema
      tags:
      - wiki-schema
...